% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editableInput.R
\name{editableInput}
\alias{editableInput}
\title{Create an in-place editable input control}
\usage{
editableInput(inputId, type, value = "", options = list())
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{type}{Type of input. One of "text", "textarea", "select", "date",
"checklist"}

\item{value}{Initial value.}

\item{options}{A list of general and input-type-dependent options for
\href{http://github.com/vitalets/x-editable}{X-editable}. Details
\href{http://vitalets.github.io/x-editable/docs.html#editable}{here}.}
}
\value{
An in-place editable input control that can be added to a UI definition.
}
\description{
Create an in-place editable input control
}
\examples{
editableInput("foo", "text", "bar")

}
