% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editableInput.R
\name{updateEditableInput}
\alias{updateEditableInput}
\title{Change the value of an in-place editable  input on the client}
\usage{
updateEditableInput(session, inputId, value = NULL)
}
\arguments{
\item{session}{The \code{session} object passed to function given to shinyServer.}

\item{inputId}{The id of the input object.}

\item{value}{The value to set for the input object.}
}
\description{
Change the value of an in-place editable  input on the client
}
\examples{
## Only run examples in interactive R sessions
if (interactive()) {

ui <- fluidPage(
  editableInput("foo", "text", ""),
  actionButton("update", "Update"),
  verbatimTextOutput("bar")
)

server <- function(input, output, session) {
  output$bar <- renderPrint(input$foo)
  observeEvent(input$update, {
    updateEditableInput(session, "foo", sample(month.abb, 1))
  })
}

shinyApp(ui, server)

}
}
